Extract sprite and custom content endpoints to a separate files. Update constants references
