+++ public/scripts/extensions/expressions/index.js
@@ -1249,7 +1249,7 @@ async function onClickExpressionUpload(event) {
         formData.append('label', id);
         formData.append('avatar', file);
 
-        await handleFileUpload('/upload_sprite', formData);
+        await handleFileUpload('/api/sprites/upload', formData);
 
         // Reset the input
         e.target.form.reset();
@@ -1355,7 +1355,7 @@ async function onClickExpressionUploadPackButton() {
         formData.append('name', name);
         formData.append('avatar', file);
 
-        const { count } = await handleFileUpload('/upload_sprite_pack', formData);
+        const { count } = await handleFileUpload('/api/sprites/upload-zip', formData);
         toastr.success(`Uploaded ${count} image(s) for ${name}`);
 
         // Reset the input
@@ -1382,7 +1382,7 @@ async function onClickExpressionDelete(event) {
     const name = $('#image_list').data('name');
 
     try {
-        await fetch('/delete_sprite', {
+        await fetch('/api/sprites/delete', {
             method: 'POST',
             headers: getRequestHeaders(),
             body: JSON.stringify({ name, label: id }),
